services/bot_gateway/requirements.txt
aiogram==3.12.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.8.2

